+++ packages/client/modules/teamDashboard/components/TeamSettings/TeamSettings.tsx
@@ -77,12 +77,12 @@ const TeamSettings = (props: Props) => {
   return (
     <TeamSettingsLayout>
       <PanelsLayout>
-        {tier === 'personal' && (
+        {tier === 'starter' && (
           <Panel>
             <StyledRow>
-              <div>{'This team is currently on a personal plan.'}</div>
+              <div>{'This team is currently on a starter plan.'}</div>
               <PrimaryButton onClick={() => history.push(`/me/organizations/${orgId}`)}>
-                {`Upgrade Team to ${TierLabel.PRO}`}
+                {`Upgrade Team to ${TierLabel.TEAM}`}
               </PrimaryButton>
             </StyledRow>
           </Panel>